chore: move redirects parsing code
